Не работает установка смещения по напряжению

+++ sources/Device/src/FPGA/MathFPGA.cpp
@@ -130,9 +130,9 @@ float MathFPGA::VoltageCursor(float shiftCurU, Range::E range, int16 rShift)
 }
 
 
-int MathFPGA::RShift2Rel(float rShiftAbs, Range::E range)
+int16 MathFPGA::RShift2Rel(float rShiftAbs, Range::E range)
 {
-    int retValue = RShift::ZERO + static_cast<int>(rShiftAbs / absStepRShift[range]);
+    int16 retValue = RShift::ZERO + static_cast<int16>(rShiftAbs / absStepRShift[range]);
 
     if (retValue < RShift::MIN)
     {
@@ -504,7 +504,7 @@ static float MaxVoltageOnScreen(Range::E range)
 }
 
 
-float MathFPGA::RShift2Abs(int rShift, Range::E range)
+float MathFPGA::RShift2Abs(int16 rShift, Range::E range)
 {
     return -(RShift::ZERO - static_cast<int>(rShift)) * absStepRShift[range];
 }
